# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Flowra API
  description: |
    Chat System with Task Management API.

    This API provides comprehensive functionality for:
    - User authentication (Keycloak SSO integration)
    - Workspace management with role-based access control
    - Real-time chat with participants management
    - Task management with kanban-style workflows
    - Message handling with attachments and reactions
    - Notification system
    - WebSocket real-time communication

    ## Authentication

    All endpoints (except `/auth/login` and `/auth/register`) require JWT authentication.
    Include the JWT token in the `Authorization` header:

    ```
    Authorization: Bearer <token>
    ```

    ## Rate Limiting

    - Default: 100 requests per minute per user
    - Auth endpoints: 10 requests per minute per IP
    - WebSocket: 60 messages per minute per connection

    ## Error Handling

    All errors follow a consistent format with `code` and `message` fields.
  version: 1.0.0
  contact:
    name: Flowra Development Team
    email: dev@flowra.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://api.flowra.com/api/v1
    description: Production server
security:
  - bearerAuth: []
...
paths:
...
  /workspaces/{workspace_id}/chats/{chat_id}/messages:
    get:
      tags:
        - Messages
      summary: List messages in chat
      description: Returns messages in a chat with pagination
      operationId: listMessages
      parameters:
        - $ref: "#/components/parameters/WorkspaceIdPath"
        - $ref: "#/components/parameters/ChatIdPath"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - name: before
          in: query
          description: Get messages before this message ID (cursor-based pagination)
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageListResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"

    post:
      tags:
        - Messages
      summary: Send a message
      description: Sends a new message to the chat
      operationId: sendMessage
      parameters:
        - $ref: "#/components/parameters/WorkspaceIdPath"
        - $ref: "#/components/parameters/ChatIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageRequest"
            example:
              content: "Hello, team! Let's discuss the project."
              reply_to_id: null
      responses:
        "201":
          description: Message sent successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
  /messages/{message_id}:
    put:
      tags:
        - Messages
      summary: Edit a message
      description: Edits a message. Only the author can edit their message.
      operationId: editMessage
      parameters:
        - $ref: "#/components/parameters/MessageIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditMessageRequest"
            example:
              content: "Updated message content"
      responses:
        "200":
          description: Message edited successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          description: Only the message author can edit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          $ref: "#/components/responses/NotFoundError"

    delete:
      tags:
        - Messages
      summary: Delete a message
      description: Soft-deletes a message. Only the author or chat admin can delete.
      operationId: deleteMessage
      parameters:
        - $ref: "#/components/parameters/MessageIdPath"
      responses:
        "204":
          description: Message deleted successfully
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"

  # ============================================
  # Task Endpoints
  # ============================================
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login

  # ============================================
  # Parameters
  # ============================================
  parameters:
    WorkspaceIdPath:
      name: workspace_id
      in: path
      required: true
      description: Workspace ID
      schema:
        type: string
        format: uuid
    ChatIdPath:
      name: chat_id
      in: path
      required: true
      description: Chat ID
      schema:
        type: string
        format: uuid
...
    MessageIdPath:
      name: message_id
      in: path
      required: true
      description: Message ID
      schema:
        type: string
        format: uuid
...
    Limit:
      name: limit
      in: query
      description: Maximum number of results to return
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Offset:
      name: offset
      in: query
      description: Number of results to skip
      schema:
        type: integer
        minimum: 0
        default: 0

  # ============================================
  # Responses
  # ============================================
  responses:
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            success: false
            error:
              code: "UNAUTHORIZED"
              message: "Authentication required"
    ForbiddenError:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            success: false
            error:
              code: "FORBIDDEN"
              message: "You don't have permission to perform this action"
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            success: false
            error:
              code: "NOT_FOUND"
              message: "Resource not found"
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            success: false
            error:
              code: "VALIDATION_ERROR"
              message: "Invalid request data"
...
  schemas:
...
    Error:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Error code
              example: "VALIDATION_ERROR"
            message:
              type: string
              description: Human-readable error message
              example: "Invalid request data"
            details:
              type: object
              description: Additional error details
              additionalProperties:
                type: string

    # Authentication schemas
...
    SendMessageRequest:
      type: object
      required:
        - content
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 10000
        reply_to_id:
          type: string
          format: uuid
          description: ID of the message being replied to
    EditMessageRequest:
      type: object
      required:
        - content
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 10000
    MessageResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            id:
              type: string
              format: uuid
            chat_id:
              type: string
              format: uuid
            sender_id:
              type: string
              format: uuid
            content:
              type: string
            reply_to_id:
              type: string
              format: uuid
            created_at:
              type: string
              format: date-time
            edited_at:
              type: string
              format: date-time
            is_deleted:
              type: boolean
            attachments:
              type: array
              items:
                type: object
                properties:
                  file_id:
                    type: string
                    format: uuid
                  file_name:
                    type: string
                  file_size:
                    type: integer
                  mime_type:
                    type: string
            reactions:
              type: array
              items:
                type: object
                properties:
                  emoji:
                    type: string
                  users:
                    type: array
                    items:
                      type: string
                      format: uuid
                  count:
                    type: integer
    MessageListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            messages:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  chat_id:
                    type: string
                    format: uuid
                  sender_id:
                    type: string
                    format: uuid
                  content:
                    type: string
                  reply_to_id:
                    type: string
                    format: uuid
                  created_at:
                    type: string
                    format: date-time
                  edited_at:
                    type: string
                    format: date-time
                  is_deleted:
                    type: boolean
            has_more:
              type: boolean
            next_cursor:
              type: string

    # Task schemas
...